refactor(app): move four impl QuantickApp groups into sibling modules
`app.rs` is the first entry in the size baseline and the one file this
sprint's merges left the same size or larger: 9,181 production lines at
`bc39248`, 9,239 at `2254039`. Every mission so far extracted from around
it. Four groups that nothing else in the window needs to see move out, into
siblings under `crates/app/src/app/`, the shape `app/layout_wiring.rs`
already has:

  demo_hooks.rs        the harness demo appliers, and the nine DEMO_*
                       constants only they read
  drawing_input.rs     handle_drawing_keys and apply_drawing_chrome
  health.rs            maybe_emit_summary, status_model, SUMMARY_INTERVAL
  workspace_restore.rs restore_workspace

Bodies are unchanged. The only edits to a moved line are the twelve
`fn` -> `pub(super) fn` widenings a caller left behind needs; the four
methods called only by moved code stay private. `pub(super)` reaches
`app::tests` too, so no test gains visibility it did not already have.

No hook declaration moves. Every `std::env::var` of an `app.rs`-declared
hook is in `new_with_workspace` or `harness.rs`; the moved regions name
`QUANTICK_*` only in doc comments and log messages, which the registry
guard does not count as reads. `hooks::OWNERS` and the generated registry
are untouched.

`DrawingsDemo` reached the test files through `app.rs`'s own imports and is
now bound in `app/tests/mod.rs` beside the other test-only binding, so the
one `use super::*` per file still carries it.

app.rs: 9,241 -> 7,501 lines.
