+++ crates/app/src/app.rs
@@ -18,16 +18,17 @@ use eframe::egui;
 
 use crate::canvas_layout::{MAX_CANVAS_PANES, PaneIdAllocator};
 
+mod demo_hooks;
+mod drawing_input;
+mod health;
 mod layout_wiring;
+mod workspace_restore;
 use crate::chart_layers::{self, ChartLayer};
 use crate::config::AppConfig;
 use crate::dock::{Dock, DockEnv, DockTab};
 use crate::drawings::{self, DeleteOutcome, DrawingAuthor};
 use crate::feed_notice;
-use crate::harness::{
-    ContextMenuPane, DrawingDraft, DrawingsDemo, FrvpDemo, Harness, ScriptedMenu, StrategyDemoMode,
-    VenueHistoryDemo,
-};
+use crate::harness::{ContextMenuPane, Harness, ScriptedMenu};
 use crate::indicator_legend;
 use crate::indicator_panel::{self, SettingsDialog, SettingsOutcome};
 use crate::indicator_worker::{IndicatorCommand, IndicatorEvent, IndicatorSource, SlotId};
@@ -64,46 +65,6 @@ const FIRST_TAB_ID: u64 = 0;
 /// objects across. Close to what a default viewport shows, so every object
 /// lands on screen — a demo the camera cannot see proves nothing.
 const DEMO_VISIBLE_SLOTS: usize = 90;
-/// How many demo objects wide the visible window is — the reciprocal of how
-/// far a multi-anchor object reaches. Four keeps a rectangle big enough to
-/// read while still leaving the tools distinguishable from each other.
-const DEMO_SPANS_PER_WINDOW: usize = 4;
-/// How far apart, as a fraction of the visible price band, the demo places
-/// the successive anchors of one object and the successive rows of objects.
-/// Both are small enough that the widest object still lands inside the band
-/// the chart is showing.
-const DEMO_ANCHOR_BAND_STEP: f64 = 0.12;
-const DEMO_ROW_BAND_STEP: f64 = 0.22;
-/// Points the demo hook gives a freehand tool, which declares no anchor
-/// count of its own. Enough to read as a path rather than as a line.
-const DEMO_FREEHAND_POINTS: usize = 4;
-/// How much of the visible window and of the visible price band the
-/// `QUANTICK_DRAWING_DRAFT` hook's anchors span. Wide enough that the
-/// half-made object reads at a glance, and centred, so the segment its
-/// anchors describe passes through the parked pointer at the chart's middle.
-///
-/// Much wider than it is tall, because a trend line a trader would actually
-/// draw runs *along* the tape. A near-vertical draft photographs the
-/// mechanism but nothing about whether the shape reads, which is what a QA
-/// reference image is for.
-const DEMO_DRAFT_SPAN: f32 = 0.6;
-const DEMO_DRAFT_BAND_SPAN: f64 = 0.12;
-/// Where the parked pointer stands when a single anchor is down, as a
-/// fraction of the chart from its centre.
-///
-/// A lone anchor sits at that centre, so parking the pointer there too gives
-/// a rubber band of no length — an invisible draft, which is the one thing
-/// this hook exists not to photograph. Offset, there is a line to see, and it
-/// is a *sloped* one, which is what makes the levelled state worth its own
-/// capture.
-const DEMO_DRAFT_POINTER_OFFSET: egui::Vec2 = egui::vec2(0.2, -0.15);
-/// The band to spread across before the pane has an auto-range to read (no
-/// bars yet): a fraction of price, since there is nothing better to ask.
-const DEMO_FALLBACK_BAND_FRACTION: f64 = 0.004;
-/// How far before the loaded history the re-cut demo anchors its off-series
-/// mark. Any distance the tab cannot possibly hold would do; an hour is
-/// unambiguous at every timeframe the chart offers.
-const DEMO_OFF_SERIES_LEAD_MS: i64 = 3_600_000;
 /// The natives `QUANTICK_INDICATORS_AUTOSTART` opens with: the overlay and
 /// the pane, so a scripted run photographs both shapes. Named by catalog id
 /// rather than "all of them", because the hook's contract is a fixed,
@@ -200,9 +161,6 @@ fn indicator_preview_area(
         .map(|tab| tab.pane(target.side))
         .and_then(|pane| pane.last_chart_area)
 }
-
-/// How often the perf summary is logged (not every frame).
-const SUMMARY_INTERVAL: Duration = Duration::from_secs(2);
 
 /// Format the forming bar's countdown, e.g. `37/50 ticks`.
 ///
@@ -2084,135 +2042,6 @@ impl QuantickApp {
         outcome
     }
 
-    /// The `QUANTICK_CONTROL_EVIDENCE` hook: capture one evidence bundle
-    /// through the very read a connected client calls.
-    ///
-    /// The value is a comma-separated list of tokens: `all` (or `1`) means
-    /// every registered scope the configured grant already reaches,
-    /// `screenshot` asks for the window to be rasterised as well, and
-    /// anything else is a snapshot scope ID. The manifest is logged, so a
-    /// scripted validation run reads what the bundle covered — and what it
-    /// did not — without a client on the socket.
-    ///
-    /// A capture that asked for an image waits for it: the window is asked to
-    /// rasterise and the hook takes the next frame, giving up after
-    /// [`crate::harness::CONTROL_EVIDENCE_HOOK_FRAMES`] rather than hanging a capture run on a
-    /// surface that never presents.
-    fn apply_control_evidence_hook(&mut self, ctx: &egui::Context) {
-        if self.pending_control_evidence.is_none() {
-            return;
-        }
-        // Access is taken *before* the request is, so a frame that finds it
-        // borrowed leaves the hook pending rather than dropping it silently.
-        let Some(mut access) = self.control_access.take() else {
-            return;
-        };
-        let request = self
-            .pending_control_evidence
-            .take()
-            .expect("the hook was pending one line above");
-        let mut wants_screenshot = false;
-        // A set, not a list: `all,scene.controls` is a reasonable thing to
-        // type, and the capability refuses a scope named twice, so the tokens
-        // are folded rather than concatenated.
-        let mut scopes = std::collections::BTreeSet::new();
-        for token in request.split(',').map(str::trim).filter(|t| !t.is_empty()) {
-            match token {
-                "screenshot" => wants_screenshot = true,
-                "all" | "1" => scopes.extend(
-                    access
-                        .readable_scopes()
-                        .into_iter()
-                        .map(|scope| scope.to_string()),
-                ),
-                scope => {
-                    scopes.insert(scope.to_owned());
-                }
-            }
-        }
-        if scopes.is_empty() {
-            scopes.extend(
-                access
-                    .readable_scopes()
-                    .into_iter()
-                    .map(|scope| scope.to_string()),
-            );
-        }
-        // The scope is checked before anything is armed, not after. Arming is
-        // what eventually raises the screenshot notice, and telling the trader
-        // their window was captured on the way to refusing the capture would
-        // make the one indicator `visual-qa` asserts on say something untrue.
-        if wants_screenshot && !access.grants_screenshot() {
-            tracing::warn!(
-                target: "quantick::control",
-                event_code = "CONTROL_EVIDENCE_HOOK_SCREENSHOT_NOT_GRANTED",
-                "QUANTICK_CONTROL_EVIDENCE asked for an image without observe.screenshot; \
-                 capturing without one"
-            );
-            wants_screenshot = false;
-        }
-        if wants_screenshot {
-            // Harvests as well as arms: the frame service that normally takes
-            // the pixels runs only while the gateway is enabled, and this hook
-            // is meant to work without a client on the socket.
-            access.service_screenshot(self, ctx);
-        }
-        if wants_screenshot && !access.has_screenshot() {
-            if self.harness.evidence_frame_waited() {
-                // Every waiting frame asks for the next one. Without this a
-                // quiescent window — a paused replay, no feed, exactly the
-                // headless validation run the hook exists for — would never
-                // repaint, the counter would never advance, and the hook would
-                // neither complete nor give up.
-                ctx.request_repaint();
-                self.pending_control_evidence = Some(request);
-                self.control_access = Some(access);
-                return;
-            }
-            tracing::warn!(
-                target: "quantick::control",
-                event_code = "CONTROL_EVIDENCE_HOOK_GAVE_UP_ON_IMAGE",
-                frames = crate::harness::CONTROL_EVIDENCE_HOOK_FRAMES,
-                "the window never delivered a frame to rasterise; capturing without one"
-            );
-            // The request stays true on purpose. Clearing it would send
-            // `screenshot: false`, and the bundle would then record
-            // `screenshot/not_requested` — a lie about a run that asked for a
-            // picture and waited two seconds for one. Left true, the capture
-            // finds no image and records `frame_not_delivered`, which is what
-            // actually happened.
-        }
-        let outcome = access.invoke_local_read(
-            self,
-            "evidence.capture",
-            serde_json::json!({ "scopes": scopes, "screenshot": wants_screenshot }),
-        );
-        self.control_access = Some(access);
-        match outcome {
-            Ok(manifest) => tracing::info!(
-                target: "quantick::control",
-                event_code = "CONTROL_EVIDENCE_CAPTURED",
-                evidence_id = %manifest["evidence_id"].as_str().unwrap_or_default(),
-                content_digest = %manifest["content_digest"].as_str().unwrap_or_default(),
-                encoded_bytes = %manifest["encoded_bytes"].as_str().unwrap_or_default(),
-                chunk_count = manifest["chunk_count"].as_u64().unwrap_or_default(),
-                captured_scopes = manifest["source_scopes"].as_array().map_or(0, Vec::len),
-                omitted_scopes = manifest["coverage"]["omitted_scopes"].as_array().map_or(0, Vec::len),
-                not_captured = manifest["coverage"]["not_captured"].as_array().map_or(0, Vec::len),
-                unavailable_fields = manifest["coverage"]["unavailable_fields"].as_array().map_or(0, Vec::len),
-                screenshot = !manifest["screenshot"].is_null(),
-                "an evidence bundle was captured through the control plane"
-            ),
-            Err(error) => tracing::warn!(
-                target: "quantick::control",
-                event_code = "CONTROL_EVIDENCE_HOOK_REFUSED",
-                error_code = %error.code,
-                error = %error.message,
-                "QUANTICK_CONTROL_EVIDENCE could not capture a bundle"
-            ),
-        }
-    }
-
     /// A launch hook's action, with its failure reported where a scripted run
     /// will see it: the hook is fire-and-forget, so nothing else would.
     fn run_hook_action(&mut self, capability_id: &str, input: serde_json::Value) {
@@ -3980,146 +3809,6 @@ impl QuantickApp {
         (tabs, chrome)
     }
 
-    /// Open the saved workspace over the configured defaults.
-    ///
-    /// The first tab already exists and is already streaming the market
-    /// `main` picked from this same workspace, so it is *arranged* here rather
-    /// than opened; the rest are opened outright, each on its own feed. A tab
-    /// carries its bar rule explicitly (see [`Self::adopt_tab`]) — inheriting
-    /// would replace what the user saved with what the tab beside it happens
-    /// to show.
-    ///
-    /// `save_on_exit` is taken from the file even when the file has no tabs:
-    /// a trader who switched autosave off and then reset their layout must not
-    /// find it switched back on at the next launch.
-    fn restore_workspace(&mut self, workspace: ui_state::Workspace) {
-        self.workspace.session_mut().adopt(
-            workspace.save_on_exit,
-            workspace.saved.clone(),
-            workspace.recent_workspaces.clone(),
-        );
-        self.refresh_recent_workspaces();
-        // One stat at boot, so the Reset entry can gate on a field instead of
-        // the filesystem for the rest of the session. A file with no tabs
-        // still counts: it carries the autosave setting, and Reset is how the
-        // trader gets rid of it.
-        let on_disk = self.workspace.ui_state_path().exists();
-        self.workspace.session_mut().set_saved(on_disk);
-        // Outside the chrome block deliberately: the stars belong to the file,
-        // not to the arrangement, so a workspace with nothing else in it still
-        // hands the rail back its pinned section.
-        //
-        // An empty list is silence, not an instruction. The format cannot tell
-        // "the trader starred nothing" from "this file predates the field" or
-        // "this bundle was written by an install that never saved a cockpit",
-        // and this same function restores an *imported* workspace mid-session —
-        // where emptying the rail on silence would throw away a curated rail on
-        // the strength of a key that was never written. Unstarring the last
-        // tool is not lost by this: that click writes the empty list itself,
-        // and the rail it would be restored onto is already empty.
-        if !workspace.favorite_tools.is_empty() {
-            let unknown = self.toolrail.set_favorites(&workspace.favorite_tools);
-            if !unknown.is_empty() {
-                // Said out loud because the next star click writes the pruned
-                // list back over the file: this is the only moment the id
-                // still exists anywhere.
-                tracing::warn!(
-                    target: "quantick::app",
-                    schema_version = 1_u8,
-                    event_code = "TOOL_FAVORITE_DROPPED",
-                    tools = %unknown.join(","),
-                    action = "no_such_drawing_tool",
-                    "a starred tool this build does not offer was dropped from the rail"
-                );
-            }
-        }
-        if let Some(chrome) = &workspace.chrome {
-            self.restore_chrome(chrome);
-        }
-        if workspace.is_empty() {
-            return;
-        }
-        // Whether tab zero is already the workspace's first market.
-        //
-        // At startup it is: `main` read that market out of this very file and
-        // spawned the window on it, so the tab exists and only its bar rule
-        // can differ. Mid-session — a workspace file being opened — it is
-        // whatever the trader was looking at, and adopting it would leave the
-        // strip holding one market with another's name. Then every tab is
-        // opened outright and the ones that were there are closed after, so
-        // the strip is *replaced* rather than grown.
-        let adopt_first = self.tabs.first().is_some_and(|tab| tab.id == FIRST_TAB_ID)
-            && self.tabs.len() == 1
-            && self.tabs[0].feed_id == workspace.tabs[0].feed
-            && self.tabs[0].symbol == workspace.tabs[0].symbol;
-        let stale: Vec<u64> = if adopt_first {
-            Vec::new()
-        } else {
-            self.tabs.iter().map(|tab| tab.id).collect()
-        };
-        for (index, saved) in workspace.tabs.iter().enumerate() {
-            // `restore` has already dropped anything unparseable, so a spec
-            // reaching here is one a control could have produced.
-            let flow = BarSpec::parse(&saved.flow_bars).ok();
-            if index == 0 && adopt_first {
-                // Tab zero is the one `main` spawned. Its market matches this
-                // entry (that is where `main` read it from), so only its bar
-                // rule can still differ — `main` prefers a feed's declared
-                // `default_bars` when the workspace names none.
-                if let Some(spec) = flow {
-                    self.tabs[0].flow_pane.set_spec(spec);
-                }
-            } else {
-                self.open_tab(saved.feed.clone(), saved.symbol.clone(), flow);
-            }
-            let context_intervals =
-                saved_context_intervals(&saved.context_bars, saved.time_bars.as_deref());
-            let focus = saved.focus.map(|focus| focus.to_side(saved.focus_slot));
-            // `open_tab` activates what it opened, so the tab just arranged is
-            // always the last one — index zero on the first pass.
-            let target = if index == 0 && adopt_first {
-                0
-            } else {
-                self.tabs.len() - 1
-            };
-            self.tabs[target].restore_canvas(
-                CanvasLayout::from(saved.layout),
-                saved.split_fraction,
-                saved.context_collapsed,
-                focus,
-                &context_intervals,
-                LegendFold {
-                    flow: saved.flow_legend_collapsed,
-                    time: saved.time_legend_collapsed,
-                },
-            );
-            self.tabs[target].set_opening_layouts(saved.flow_layout, &saved.context_layouts);
-        }
-        // The markets that were on screen before this workspace was opened.
-        // Closed last, so the strip is never empty in between and `close_tab`
-        // — which refuses to close the only tab — always has the restored
-        // ones to keep. Each closes through its own path, so a simulated
-        // position ends in a journaled flatten rather than vanishing.
-        for id in stale {
-            if let Some(index) = self.tabs.iter().position(|tab| tab.id == id) {
-                self.close_tab(index);
-            }
-        }
-        self.active_tab = workspace.active_tab.min(self.tabs.len() - 1);
-        let config = self.config.clone();
-        self.active_tab_mut().refresh_chip_label(&config);
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "UI_STATE_RESTORED",
-            path = %self.workspace.ui_state_path().display(),
-            tabs = self.tabs.len(),
-            active = self.active_tab,
-            save_on_exit = self.workspace.session().save_on_exit(),
-            "workspace restored"
-        );
-    }
-
     /// Ask the operating system where to put a workspace file, off the UI
     /// thread. One dialog at a time, the trades-folder picker's own pattern.
     ///
@@ -5062,309 +4751,6 @@ impl QuantickApp {
     fn note_workspace(&mut self, message: String) {
         self.surfaces.toast.note(message, Instant::now());
     }
-
-    /// Periodically log a perf summary and warn on threshold breaches.
-    fn maybe_emit_summary(&mut self, now: Instant, ctx: &egui::Context) {
-        let elapsed = now - self.last_summary;
-        if elapsed < SUMMARY_INTERVAL {
-            return;
-        }
-        // The window's own geometry, because a chart that lays out wider than
-        // the surface it is painted on loses its right edge — the toolbar's
-        // layer group, the price axis, the live strip and the dock all live
-        // there — and nothing else in this line would say so.
-        //
-        // `client_px` is the platform's *own* answer, not `screen * scale`:
-        // that product is algebraically the same number as `screen_pt` beside
-        // it and could never contradict anything, which would make the whole
-        // block decoration. Two independent readings, so the line can be read
-        // for whether they agree — and by the time it is written a correction
-        // may already have restored that agreement, which `WINDOW_SCALE_CORRECTED`
-        // is the record of.
-        // Read outside the `input` closure: egui holds one lock for the whole
-        // of it, and reaching back into the context from inside deadlocks.
-        let zoom = ctx.zoom_factor();
-        let client = self
-            .surface
-            .as_ref()
-            .and_then(window_scale::SurfaceProbe::client_size_px);
-        let (screen, scale, native_scale) = ctx.input(|input| {
-            (
-                input.screen_rect.size(),
-                input.pixels_per_point(),
-                input.viewport().native_pixels_per_point,
-            )
-        });
-        let rate = self.trades_since_summary as f64 / elapsed.as_secs_f64();
-        let lag = self.active_tab().trade_arrival_ms();
-        let avg = self.frames.avg_ms().unwrap_or(0.0);
-        let cpu_avg = self.cpu_frames.avg_ms().unwrap_or(0.0);
-        let worst = self.frames.worst_ms().unwrap_or(0.0);
-        let fps = self.frames.fps().unwrap_or(0.0);
-        let book = self.active_tab_mut().tape_mut().health();
-        let book_lag = book.arrival_latency_ms;
-        let book_rate = book.depth_updates_since_summary as f64 / elapsed.as_secs_f64();
-        let book_queue_len = self.active_tab().book_events.len();
-        let candle_preset =
-            CandlePreset::detect(&self.style.candles).map_or("custom", CandlePreset::log_value);
-
-        tracing::info!(
-            target: "quantick::app",
-            schema_version = 1_u8,
-            event_code = "APP_HEALTH_SUMMARY",
-            // Frames and the trade rate are the window's; every market figure
-            // below is the *active* tab's, which is what is on screen.
-            tabs = self.tabs.len(),
-            tab = self.active_tab().id,
-            fps = fps as i64,
-            frame_avg_ms = avg,
-            frame_cpu_ms = cpu_avg,
-            frame_worst_ms = worst,
-            feed_arrival_ms = lag,
-            trades_per_s = rate,
-            live_trades = self.active_tab().live_trades,
-            bar_spec = self.active_tab().flow_pane.state.spec().summary(),
-            // Both facts the tape states about its own prices — the grid they
-            // land on and the magnitude they land at — and the row width the
-            // ladder ended up drawing from them. A validation run reads all
-            // three: the first two are the sizing rule's whole input, so
-            // without them a run can see that rows are 1.00 and not why, and
-            // cannot tell "the chart has not been told a tick yet" from "the
-            // chart is drawing rows finer than the instrument can trade at".
-            tape_price_step = self
-                .active_tab()
-                .flow_pane
-                .state
-                .tape_price_step()
-                .map_or_else(|| "unknown".to_owned(), |step| step.to_string()),
-            tape_reference_price = self
-                .active_tab()
-                .flow_pane
-                .state
-                .tape_reference_price()
-                .map_or_else(|| "unknown".to_owned(), |price| price.to_string()),
-            footprint_rows = %self.active_tab().flow_pane.state.footprint_group(),
-            canvas_layout = ?self.active_tab().layout,
-            screen_pt_w = screen.x,
-            screen_pt_h = screen.y,
-            client_px_w = client.map(|size| size.x),
-            client_px_h = client.map(|size| size.y),
-            scale = scale,
-            native_scale = native_scale,
-            zoom_factor = zoom,
-            time_pane_spec = self.active_tab().time_pane().map(|pane| pane.state.spec().summary()),
-            time_pane_count = self.active_tab().time_panes.len(),
-            // Drawings are a per-frame, O(objects) paint cost, and the shared
-            // ones are additionally reprojected on every other pane of the
-            // tab. Counting them here is what lets a frame-cost reading be
-            // attributed instead of guessed — and it is the only way a
-            // headless run can prove the drawing overlay is populated at all.
-            drawings = self
-                .active_tab()
-                .panes()
-                .map(|(pane, _)| pane.drawings.items().len())
-                .sum::<usize>(),
-            shared_drawings = self
-                .active_tab()
-                .panes()
-                .map(|(pane, _)| pane.drawings.shared_count())
-                .sum::<usize>(),
-            book_enabled = book.enabled,
-            book_status = book.status,
-            book_generation = book.generation,
-            book_last_update_id = book.last_update_id,
-            book_last_event_ms = book.last_event_ms,
-            book_snapshot_observed_ms = book.last_snapshot_observed_ms,
-            book_arrival_ms = book_lag,
-            // How far the newest print sits behind the instant the lane calls
-            // now. It is the pixel gap between the last bubble and the tape's
-            // right edge, in milliseconds: a number, so "the bubbles are
-            // trailing" can be measured rather than argued about.
-            //
-            // A distance between two *venue* clocks, not staleness against
-            // this machine's. A dead session stops both, so this figure
-            // freezes rather than growing — `feed_arrival_ms` above is the one
-            // that answers "is anything still arriving".
-            tape_age_ms = book.tape_age.map(|age| match age {
-                quantick_orderflow::TapeAge::Behind(ms) | quantick_orderflow::TapeAge::NothingYet(ms) => ms,
-            }),
-            tape_age_kind = book.tape_age.map(|age| match age {
-                quantick_orderflow::TapeAge::Behind(_) => "behind",
-                quantick_orderflow::TapeAge::NothingYet(_) => "nothing_yet",
-            }),
-            book_updates_per_s = book_rate,
-            book_updates_total = book.depth_updates,
-            book_queue_len,
-            book_channel_closed = self.active_tab().book_channel_closed_reported,
-            book_bid_levels = book.bid_levels,
-            book_ask_levels = book.ask_levels,
-            heatmap_active_levels = book.active_levels,
-            heatmap_archived_runs = book.archived_runs,
-            aggression_count = book.aggression_count,
-            heatmap_history_bytes = book.history_bytes,
-            heatmap_cells = book.projection_cells,
-            heatmap_aggressions = book.projection_aggressions,
-            heatmap_liquidity_events = book.projection_liquidity_events,
-            heatmap_effective_grouping = %book.effective_grouping,
-            heatmap_effective_grouping_multiple = book.effective_grouping_multiple,
-            heatmap_dropped_cells = book.dropped_cells,
-            heatmap_folded_aggressions = book.folded_aggressions,
-            heatmap_dropped_liquidity_events = book.dropped_liquidity_events,
-            heatmap_projection_ms = book.projection_ms,
-            heatmap_live_ms = book.live_ms,
-            heatmap_projection_builds = book.projection_builds,
-            heatmap_projection_cache_hits = book.projection_cache_hits,
-            heatmap_config_revision = book.config_revision,
-            heatmap_snapshots = book.snapshots,
-            heatmap_gaps = book.gaps,
-            candle_style_revision = self.style_revision,
-            candle_preset,
-            candle_body_mode = ?self.style.candles.body_mode,
-            candle_fill_opacity = self.style.candles.fill_opacity,
-            candle_outline_opacity = self.style.candles.outline_opacity,
-            candle_outline_width_px = self.style.candles.outline_width,
-            chart_background_enabled = self.style.canvas.background_enabled,
-            chart_grid_enabled = self.style.canvas.grid_enabled,
-            replay_active = self.active_tab().replay.is_some(),
-            replay_speed = self.active_tab().replay.as_ref().map(|r| r.status.speed()),
-            replay_playing = self.active_tab().replay.as_ref().map(|r| r.status.is_playing()),
-            replay_progress = self.active_tab().replay.as_ref().map(|r| r.status.progress()),
-            replay_played = self.active_tab().replay.as_ref().map(|r| r.status.played()),
-            replay_total = self.active_tab().replay.as_ref().map(|r| r.status.total()),
-            action = "observe",
-            "application health summary"
-        );
-        if avg > metrics::SLOW_FRAME_MS {
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "APP_SLOW_FRAMES",
-                frame_avg_ms = avg,
-                threshold_ms = metrics::SLOW_FRAME_MS,
-                heatmap_enabled = book.enabled,
-                heatmap_projection_ms = book.projection_ms,
-                heatmap_cells = book.projection_cells,
-                action = "inspect_render_budget",
-                "slow frames: the chart is not keeping up"
-            );
-        }
-        if let Some(l) = lag
-            && l > metrics::HIGH_LAG_MS
-        {
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "APP_HIGH_TRADE_LAG",
-                feed_lag_ms = l,
-                threshold_ms = metrics::HIGH_LAG_MS,
-                action = "inspect_trade_connection",
-                "high feed lag: trades are arriving well behind their timestamps"
-            );
-        }
-        if let Some(l) = book_lag
-            && book.enabled
-            && l > metrics::HIGH_LAG_MS
-        {
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "HEATMAP_HIGH_ARRIVAL",
-                symbol = self.active_tab().symbol.as_str(),
-                book_arrival_ms = l,
-                threshold_ms = metrics::HIGH_LAG_MS,
-                book_status = book.status,
-                action = "inspect_depth_connection",
-                // Arrival, not age: this is how late the newest accepted
-                // depth event was when it reached us, an observation frozen
-                // at that moment. A book that stops updating keeps its last
-                // figure — the tape-age readout is what catches that.
-                "order-book events are arriving late"
-            );
-        }
-        // Losses only. Folding is the expected steady state on a busy tape and
-        // loses nothing — warning about it would tell an operator (and the
-        // planned assistant reading these events) to go fix something that is
-        // not broken. The fold count still rides in the info summary above.
-        if book.dropped_cells > 0 || book.dropped_liquidity_events > 0 {
-            tracing::warn!(
-                target: "quantick::app",
-                schema_version = 1_u8,
-                event_code = "HEATMAP_PROJECTION_CAPPED",
-                symbol = self.active_tab().symbol.as_str(),
-                dropped_cells = book.dropped_cells,
-                dropped_liquidity_events = book.dropped_liquidity_events,
-                // Not "group harder". Grouping is exactly what the trader is
-                // complaining about when marks read as one blob, and the
-                // aggression budget no longer discards anything to begin with —
-                // it folds, and says how much it folded. What is worth widening
-                // is the budget or the pane, so that is what this names.
-                action = "increase_grouping_or_reduce_retention",
-                "heatmap depth primitive cap dropped items"
-            );
-        }
-
-        self.trades_since_summary = 0;
-        self.active_tab_mut().tape_mut().reset_summary_counters();
-        self.last_summary = now;
-    }
-
-    /// Everything the status bar reports this frame.
-    ///
-    /// Provenance (venue, symbol, transport, side honesty) is the market's and
-    /// reads from the window; the content section — spec, bar counts, forming
-    /// bar, whether the view follows live — is the *focused pane's* (§11), so
-    /// the bar always describes the chart the user is working in.
-    fn status_model(&self) -> statusbar::StatusModel {
-        let pane = self.focused_pane();
-        let bars = pane.state.bars();
-        let (backfilled, live) = match pane.state.backfill_boundary() {
-            Some(boundary) => (boundary, bars.len().saturating_sub(boundary)),
-            None => (0, bars.len()),
-        };
-        let venue_bars = pane.history_prefix.len();
-        let note = self.active_tab().side_note(&self.config);
-        statusbar::StatusModel {
-            venue: if self.active_tab().replay.is_some() {
-                "recording".to_owned()
-            } else {
-                self.active_tab().feed_display_name(&self.config).to_owned()
-            },
-            symbol: self.active_tab().symbol.clone(),
-            replay: self
-                .active_tab()
-                .replay
-                .as_ref()
-                .map(|link| statusbar::ReplayFigures {
-                    speed: link.status.speed(),
-                    progress: link.status.progress(),
-                }),
-            connection: self.active_tab().feed_connection,
-            feed_arrival_ms: self.active_tab().trade_arrival_ms(),
-            feed_latency: self.active_tab().feed_latency(),
-            tape_age_ms: self.active_tab().tape_age_at(metrics::wall_clock_ms()),
-            spec_summary: pane.state.spec().summary(),
-            bar_progress: pane
-                .state
-                .progress()
-                .map(|(progress, unit)| fmt_progress(&progress, unit)),
-            venue_bars,
-            backfilled_bars: backfilled,
-            live_bars: live,
-            side_note: note.clone().map(|(label, _)| label),
-            side_detail: note.and_then(|(_, detail)| detail),
-            // Provenance follows the active tab (§11), and so does the
-            // simulated P&L: the cell speaks for the market on screen, never
-            // for a background tab's position.
-            sim_pnl: self.active_tab().paper.status_cell(),
-            follows_live: pane.viewport.follows_live(),
-            price_auto: pane.price_view.is_auto(),
-            live_trades: self.active_tab().live_trades,
-            fps: self.frames.fps(),
-            frame_avg_ms: self.frames.avg_ms(),
-            frame_cpu_ms: self.cpu_frames.avg_ms(),
-            show_perf: self.show_perf,
-        }
-    }
 }
 
 /// Opens the Market Replay browser (§10).
@@ -6165,156 +5551,6 @@ impl QuantickApp {
         }
     }
 
-    /// Keyboard grammar for drawings. Any focused widget wins: while an input
-    /// owns the keyboard, chart shortcuts stay suspended.
-    fn handle_drawing_keys(&mut self, ctx: &egui::Context, now: Instant) {
-        if ctx.memory(|memory| memory.focused().is_some()) {
-            return;
-        }
-        struct DrawingKeys {
-            escape: bool,
-            delete: bool,
-            backspace: bool,
-            undo: bool,
-            redo: bool,
-            lock: bool,
-            hide: bool,
-            duplicate: bool,
-            nudge_bars: f32,
-            nudge_px: f32,
-        }
-        let keys = ctx.input(|input| {
-            let command = input.modifiers.command;
-            let shift = input.modifiers.shift;
-            let alt = input.modifiers.alt;
-            // Shift turns a nudge into ten steps (UX spec).
-            let step = if shift { 10.0 } else { 1.0 };
-            let horizontal = f32::from(input.key_pressed(egui::Key::ArrowRight))
-                - f32::from(input.key_pressed(egui::Key::ArrowLeft));
-            let vertical = f32::from(input.key_pressed(egui::Key::ArrowUp))
-                - f32::from(input.key_pressed(egui::Key::ArrowDown));
-            DrawingKeys {
-                escape: input.key_pressed(egui::Key::Escape),
-                delete: input.key_pressed(egui::Key::Delete),
-                backspace: input.key_pressed(egui::Key::Backspace),
-                undo: command && !shift && input.key_pressed(egui::Key::Z),
-                redo: (command && input.key_pressed(egui::Key::Y))
-                    || (command && shift && input.key_pressed(egui::Key::Z)),
-                lock: alt && input.key_pressed(egui::Key::L),
-                hide: alt && input.key_pressed(egui::Key::H),
-                duplicate: command && input.key_pressed(egui::Key::D),
-                nudge_bars: horizontal * step,
-                nudge_px: vertical * step,
-            }
-        });
-        // The escape stack: rail drag → paper interaction → pending
-        // confirmation → draft → selection → Pointer, one layer per press.
-        // Paper trading's armed placement / grabbed line reads Escape here,
-        // in the single stack — what keeps one press from firing two
-        // cancels at once.
-        //
-        // The context bar's parked position is deliberately *not* a layer of
-        // this stack. It is a preference, not a gesture left half-finished:
-        // a trader who moved the bar out of the way wants it to stay out of
-        // the way, and Escape is a key they press many times an hour to drop
-        // a selection. Spending it on the parked point would undo, several
-        // times a session and without being asked, the very thing parking it
-        // was for. The way back is the grip's double-click, which is aimed at
-        // the bar and at nothing else — and, for an operator with no hand on
-        // the mouse, `ContextBar::clear_manual`, which is what that
-        // double-click calls rather than reimplements.
-        if keys.escape {
-            if self.toolrail.drag_active() {
-                // The rail consumes this Esc to abort its dock drag.
-            } else if self.active_tab_mut().paper.cancel_interaction() {
-                // An armed order placement or a grabbed order line was
-                // dropped; nothing else loses state on this press. Only the
-                // active tab can have one in flight — a background tab has
-                // no pointer over it to arm or grab with.
-            } else if self.surfaces.drawing_chrome.delete_confirm() {
-                self.surfaces.drawing_chrome.set_delete_confirm(false);
-            } else if self.inline_text_editing().is_some() {
-                // A note being typed is its own layer, and it has to be one:
-                // egui clears widget focus at the top of the frame Escape
-                // arrives on, so by the time this stack runs the editor no
-                // longer looks focused — and without a layer of its own the
-                // press fell straight through to "drop the selection",
-                // taking away the context bar for the note just written.
-                self.end_inline_text_edit();
-            } else if self.drawing_pane().drawings.draft().is_some() {
-                self.drawing_pane_mut().drawings.cancel_draft();
-                self.toolrail.arm(Tool::Pointer);
-            } else if self.drawing_pane().drawings.selected().is_some() {
-                self.drawing_pane_mut().drawings.select(None);
-            } else {
-                self.toolrail.arm(Tool::Pointer);
-            }
-        }
-        if self.drawing_pane().drawings.draft().is_some() {
-            // During placement the delete keys belong to the draft workflow:
-            // Backspace steps back one anchor.
-            if keys.backspace {
-                self.drawing_pane_mut().drawings.remove_last_draft_anchor();
-            }
-        } else if keys.delete || keys.backspace {
-            self.request_delete_selected(now);
-        }
-        if keys.undo {
-            let pane = self.drawing_pane_mut();
-            pane.drawings.undo();
-            // Undoing a rectangle's placement takes its drawing away from
-            // any armed instance without passing through the removal
-            // funnel; the sweep keeps a resting bot order from outliving
-            // its badge (redo below, and delete-all, share the risk).
-            pane.sweep_strategy_orphans();
-        }
-        if keys.redo {
-            let pane = self.drawing_pane_mut();
-            pane.drawings.redo();
-            pane.sweep_strategy_orphans();
-        }
-        if keys.lock
-            && let Some(index) = self.drawing_pane().drawings.selected()
-        {
-            let locked = self.drawing_pane().drawings.items()[index].locked;
-            self.drawing_pane_mut()
-                .drawings
-                .set_selected_locked(!locked);
-        }
-        if keys.hide
-            && let Some(index) = self.drawing_pane().drawings.selected()
-        {
-            let hidden = self.drawing_pane().drawings.items()[index].hidden;
-            self.drawing_pane_mut()
-                .drawings
-                .set_selected_hidden(!hidden);
-        }
-        if keys.duplicate {
-            self.duplicate_selected_drawing();
-        }
-        if (keys.nudge_bars != 0.0 || keys.nudge_px != 0.0)
-            && self.drawing_pane().drawings.selected().is_some()
-        {
-            // Arrows write the same honest chart coordinates a drag does:
-            // one bar per horizontal step, one pixel's worth of *that
-            // object's own axis* per vertical step. Each press lands as one
-            // undo entry. Reading the candles' scale for an object on an
-            // indicator band would nudge a CVD level by a quantity of price —
-            // a wrong number arriving through the one gesture that exists for
-            // precision. Asked of the pane that owns the mark, which for a
-            // shared one is not always the focused pane.
-            let price_per_px = self.drawing_pane().selected_value_per_px().unwrap_or(0.0);
-            self.drawing_pane_mut().drawings.begin_gesture();
-            self.drawing_pane_mut()
-                .drawings
-                .translate_selected(keys.nudge_bars, f64::from(keys.nudge_px) * price_per_px);
-            // Same rule as the drag: the instants behind the anchors move
-            // with them, or the object's shared twin stays behind.
-            self.drawing_pane_mut().retime_selected();
-            self.drawing_pane_mut().drawings.commit_gesture();
-        }
-    }
-
     /// Record one edit gesture as the single undo entry it earned, on the pane
     /// it started on.
     ///
@@ -6451,172 +5687,6 @@ impl QuantickApp {
         self.focused_pane()
             .band_label(pane.drawings.items().get(index)?)
             .chip()
-    }
-
-    /// Carry out what the drawing chrome asked for.
-    ///
-    /// One applier for all four pieces, in the order the trunk applied them
-    /// when each drew itself: the edit lands before the gesture that coalesces
-    /// it is committed, and both land before anything that can delete the
-    /// object they describe.
-    fn apply_drawing_chrome(
-        &mut self,
-        ask: crate::surfaces::drawing_chrome::DrawingChromeAsk,
-        now: Instant,
-    ) {
-        if let Some(edited) = ask.edited {
-            // Through the selection, never `items_mut`: that hatch is
-            // documented for derived-state refresh only, and a style or a
-            // note's words take part in payload equality — writing them through
-            // it would let an unrelated in-flight gesture swallow this edit.
-            if let Some(drawing) = self.drawing_pane_mut().drawings.selected_mut() {
-                *drawing = *edited;
-            }
-        }
-        if let Some(edit) = ask.commit_edit_gesture {
-            self.record_drawing_edit(edit.tab, edit.side, edit.index, edit.before);
-        }
-        if ask.toggle_selected_hidden
-            && let Some(index) = self.drawing_pane().drawings.selected()
-        {
-            let hidden = self.drawing_pane().drawings.items()[index].hidden;
-            self.drawing_pane_mut()
-                .drawings
-                .set_selected_hidden(!hidden);
-        }
-        if ask.toggle_selected_locked
-            && let Some(index) = self.drawing_pane().drawings.selected()
-        {
-            let locked = self.drawing_pane().drawings.items()[index].locked;
-            self.drawing_pane_mut()
-                .drawings
-                .set_selected_locked(!locked);
-        }
-        if ask.request_delete {
-            self.request_delete_selected(now);
-        }
-        if ask.cancel_delete {
-            // After the request, never before: a frame carrying both must end
-            // with the prompt gone, and a delete on a locked object raises it.
-            self.surfaces.drawing_chrome.set_delete_confirm(false);
-        }
-        if ask.force_delete {
-            let doomed = {
-                let pane = self.drawing_pane();
-                pane.drawings
-                    .selected()
-                    .and_then(|index| pane.drawings.items().get(index))
-                    .map(|drawing| drawing.id)
-            };
-            if self.drawing_pane_mut().drawings.delete_selected(true) == DeleteOutcome::Deleted {
-                if let Some(id) = doomed {
-                    self.drawing_pane_mut().remove_strategy_for_drawing(id);
-                }
-                self.surfaces.toast.note_with_undo("Drawing deleted.", now);
-            }
-        }
-        if ask.duplicate {
-            self.duplicate_selected_drawing();
-        }
-        if let Some(saved) = ask.saved_default {
-            // Nothing to undo: this changed a preference, not the chart.
-            self.surfaces.toast.note(saved.message(), now);
-        }
-        for write in ask.presets {
-            write.apply_to(&mut self.drawing_presets);
-        }
-        if ask.sweep_authored {
-            let removed = self.remove_every_authored_object();
-            if removed > 0 {
-                self.surfaces
-                    .toast
-                    .note_with_undo(format!("{removed} object(s) placed for you removed."), now);
-            }
-        }
-        if ask.delete_all {
-            let pane = self.drawing_pane_mut();
-            let deleted = pane.drawings.delete_all();
-            // Every armed instance just lost its drawing at once; sweep them
-            // now so no resting bot order outlives its badge.
-            pane.sweep_strategy_orphans();
-            if deleted > 0 {
-                self.surfaces
-                    .toast
-                    .note_with_undo("All drawings deleted.", now);
-            }
-        }
-        if let Some(index) = ask
-            .manager_select
-            .filter(|index| *index < self.drawing_pane().drawings.items().len())
-        {
-            self.drawing_pane_mut().drawings.select(Some(index));
-            // Centre the viewport on the object's bar span.
-            let slots = self.drawing_pane().slots();
-            if let Some(chart) = self.drawing_pane().last_chart_area {
-                let points = &self.drawing_pane().drawings.items()[index].points;
-                if !points.is_empty() {
-                    let mid =
-                        points.iter().map(|point| point.bar).sum::<f32>() / points.len() as f32;
-                    self.drawing_pane_mut()
-                        .viewport
-                        .center_on_bar(mid, chart.width(), slots);
-                }
-            }
-        }
-        // Through `get`, never `[]`: the rows were snapshotted before any of
-        // the four pieces drew, and a destructive ask applied above — delete
-        // all, the assistant sweep, a confirmed delete — can have shortened
-        // the list under an index that was valid when the row was clicked. The
-        // store's own setters are already bounds-safe; these two reads were
-        // the only raw ones left.
-        if let Some(hidden) = ask.manager_toggle_hidden.and_then(|index| {
-            Some((
-                index,
-                self.drawing_pane().drawings.items().get(index)?.hidden,
-            ))
-        }) {
-            self.drawing_pane_mut()
-                .drawings
-                .set_hidden_at(hidden.0, !hidden.1);
-        }
-        if let Some(locked) = ask.manager_toggle_locked.and_then(|index| {
-            Some((
-                index,
-                self.drawing_pane().drawings.items().get(index)?.locked,
-            ))
-        }) {
-            self.drawing_pane_mut()
-                .drawings
-                .set_locked_at(locked.0, !locked.1);
-        }
-        if let Some(index) = ask.manager_bring_to_front {
-            self.drawing_pane_mut().drawings.bring_to_front(index);
-        }
-        if let Some(index) = ask.manager_delete {
-            // The exact same command path as the inspector button and the
-            // keyboard: select, then request. Locked rows raise the same
-            // confirmation in the inspector.
-            self.drawing_pane_mut().drawings.select(Some(index));
-            self.request_delete_selected(now);
-        }
-        if ask.show_all {
-            self.drawing_pane_mut().drawings.set_all_hidden(false);
-        }
-        if ask.unlock_all {
-            self.drawing_pane_mut().drawings.set_all_locked(false);
-        }
-        if ask.place_text_note && self.place_text_note() {
-            self.surfaces.drawing_chrome.note_text_note_placed();
-        }
-        if let Some(edit) = ask.record_inline_edit {
-            self.record_drawing_edit(edit.tab, edit.side, edit.index, edit.before);
-        }
-        if ask.content_editing_changed {
-            self.sync_content_editing();
-        }
-        if self.surfaces.drawing_chrome.take_inspector_position_dirty() {
-            self.inspector_position_dirty = true;
-        }
     }
 
     /// The docked inspector.
@@ -7192,309 +6262,6 @@ impl QuantickApp {
         }
     }
 
-    /// The `QUANTICK_DRAWINGS_DEMO` hook: one of every registered drawing on
-    /// the flow pane, spread across the visible bars, the last one selected
-    /// so the inspector is on screen too.
-    ///
-    /// Waits for bars: anchors are placed on real slots, so every anchor
-    /// carries a real market time and the shared-drawing path is exercised
-    /// rather than faked. Consumed once, whether or not it placed anything on
-    /// this attempt — an env var is a request for this run, and it must never
-    /// keep re-placing objects the user then deletes.
-    fn apply_drawing_demo(&mut self) {
-        if !self.harness.drawings_demo_armed() {
-            return;
-        }
-        let pane = &mut self.active_tab_mut().flow_pane;
-        let slots = pane.slots();
-        // Enough bars for every tool to get its own stretch of chart.
-        if slots < 8 * drawings::DRAWING_TOOLS.len() {
-            return;
-        }
-        let Some(demo) = self.harness.take_drawings_demo() else {
-            return;
-        };
-        let DrawingsDemo {
-            bands,
-            shared: share,
-            select_tool,
-        } = demo;
-        if share {
-            // A shared drawing has nothing to be shared *with* on a single
-            // pane, so the hook that asks for one opens the split too — the
-            // surface under test is the projection onto the other chart.
-            self.active_tab_mut().set_layout(CanvasLayout::TimeAndFlow);
-        }
-        let pane = &mut self.active_tab_mut().flow_pane;
-        // Anchored inside the window the chart actually opens on, not at slot
-        // zero: a demo whose objects sit 300 bars off the left edge shows a
-        // screenshot of an empty chart, which is exactly the evidence this
-        // hook exists to produce. `visible` is the newest stretch, and the
-        // objects are laid across it left to right in registry order.
-        let visible = DEMO_VISIBLE_SLOTS.min(slots);
-        let first = slots - visible;
-        // Two different spacings on purpose. `stride` walks the *starts*
-        // apart so the objects are distinguishable; `span` sets how far a
-        // multi-anchor object reaches, and it has to be wide or a rectangle
-        // lands two bars across and photographs as a sliver. The objects
-        // overlap, which is fine — a QA screen wants every tool legible, not
-        // a tidy row.
-        let stride = (visible / drawings::DRAWING_TOOLS.len()).max(1);
-        let span = (visible / DEMO_SPANS_PER_WINDOW).max(2);
-        let close = pane
-            .closed_bar(slots.saturating_sub(1))
-            .and_then(|bar| rust_decimal::prelude::ToPrimitive::to_f64(&bar.close))
-            .unwrap_or(1.0);
-        // Spread the objects across the band the chart is actually showing,
-        // never across a fixed percentage of price. A tick chart's window is
-        // a few tenths of a percent tall, so a fixed ±0.2 % dropped a third
-        // of the demo below the visible range — objects placed off screen
-        // photograph as an empty chart, which is the one failure this hook
-        // exists to prevent.
-        let (centre, band) = pane
-            .last_auto_range
-            .filter(|(lo, hi)| hi > lo)
-            .map_or((close, close * DEMO_FALLBACK_BAND_FRACTION), |(lo, hi)| {
-                ((lo + hi) / 2.0, hi - lo)
-            });
-        let mut requested_selection = None;
-        for (index, tool) in drawings::DRAWING_TOOLS.into_iter().enumerate() {
-            // A freehand tool declares no anchor count, so the demo gives it
-            // a short path of its own. A stroke no screenshot can reach is a
-            // tool that ships unvalidated.
-            let anchors = if tool.freehand() {
-                DEMO_FREEHAND_POINTS
-            } else {
-                tool.required_points()
-            };
-            for anchor in 0..anchors {
-                let slot = (first + index * stride + anchor * span).min(slots.saturating_sub(1));
-                let point = drawings::ChartPoint::at_time(
-                    slot as f32 + 0.5,
-                    centre + (f64::from(anchor as i32) - 1.0) * band * DEMO_ANCHOR_BAND_STEP
-                        - (f64::from(index as i32 % 3) - 1.0) * band * DEMO_ROW_BAND_STEP,
-                    pane.slot_open_time(slot),
-                );
-                let mut completed =
-                    pane.drawings
-                        .place_with(tool, &drawings::DrawingBand::Price, point, |tool| {
-                            drawings::NewDrawing {
-                                style: tool.default_style(),
-                                payload: tool.default_payload(),
-                            }
-                        });
-                // The release, for the tool whose gesture has one.
-                if tool.freehand() && anchor + 1 == anchors {
-                    completed = pane.drawings.finish_draft();
-                }
-                // Placement selects what it completed, so this reaches the
-                // object just made — no separate index bookkeeping.
-                if completed
-                    && share
-                    && let Some(drawing) = pane.drawings.selected_mut()
-                    && drawing.shareable()
-                {
-                    drawing.scope = drawings::DrawingScope::AllCharts;
-                }
-                if completed && select_tool.as_deref() == Some(tool.id()) {
-                    requested_selection = pane.drawings.selected();
-                }
-            }
-        }
-        if let Some(index) = requested_selection {
-            pane.drawings.select(Some(index));
-            // Selecting an object the viewport is not looking at proves
-            // nothing: the handles are on screen or they are not photographed
-            // at all. Centre on the object's bar span, the object manager's
-            // own "select and centre".
-            if let Some(chart) = pane.last_chart_area {
-                let points = &pane.drawings.items()[index].points;
-                if !points.is_empty() {
-                    let mid =
-                        points.iter().map(|point| point.bar).sum::<f32>() / points.len() as f32;
-                    pane.viewport.center_on_bar(mid, chart.width(), slots);
-                }
-            }
-        }
-        if bands {
-            Self::seed_band_demo(pane, first, visible, slots);
-        }
-        self.carry_inspector_across_selection();
-        self.apply_drawing_demo_recut();
-    }
-
-    /// The `QUANTICK_DRAWING_DRAFT=<anchors>` hook: the tool armed by
-    /// `QUANTICK_DRAWING_TOOL`, part-placed, with the pointer parked where
-    /// the next anchor would go.
-    ///
-    /// The live preview of a half-made object is a surface like any other,
-    /// and for a multi-anchor tool it is the *whole* feedback of the gesture:
-    /// it is what tells the trader that a drag has fixed the trend line and a
-    /// click is owed for the width. It is also the one surface no click-free
-    /// launch could reach, because it exists only between two clicks and only
-    /// while a pointer is over the chart — so this hook does both halves.
-    ///
-    /// The anchors go down through `place_with`, the same call the click path
-    /// makes; nothing here is a parallel placement path. They straddle the
-    /// middle of the visible window so the line they draw runs through the
-    /// parked pointer — which is the exact spot the reported defect lived at,
-    /// the pointer standing on the trend line it just drew. A screenshot of
-    /// this state is a screenshot of the fix.
-    ///
-    /// Waits for bars, and is consumed once, for the same reasons
-    /// [`Self::apply_drawing_demo`] is.
-    fn apply_drawing_draft(&mut self) {
-        let Some(DrawingDraft { anchors, constrain }) = self.harness.drawing_draft() else {
-            return;
-        };
-        let Some(tool) = self.toolrail.tool().drawing_tool() else {
-            return;
-        };
-        // Bars first, and a chart rectangle to park a pointer in: both are
-        // answers only a drawn frame has.
-        let pane = &self.active_tab().flow_pane;
-        let slots = pane.slots();
-        let (Some(chart), true) = (pane.last_chart_area, slots > 0) else {
-            return;
-        };
-        self.harness.drawing_draft_staged();
-        let pane = &mut self.active_tab_mut().flow_pane;
-        // One short of the count, whatever was asked for: a draft that
-        // completed itself is not a draft, and photographs as a finished
-        // object rather than as the gesture under test.
-        let anchors = anchors.min(tool.required_points().saturating_sub(1));
-        let visible = DEMO_VISIBLE_SLOTS.min(slots);
-        let first = slots - visible;
-        let close = pane
-            .closed_bar(slots.saturating_sub(1))
-            .and_then(|bar| rust_decimal::prelude::ToPrimitive::to_f64(&bar.close))
-            .unwrap_or(1.0);
-        let (centre, band) = pane
-            .last_auto_range
-            .filter(|(lo, hi)| hi > lo)
-            .map_or((close, close * DEMO_FALLBACK_BAND_FRACTION), |(lo, hi)| {
-                ((lo + hi) / 2.0, hi - lo)
-            });
-        for anchor in 0..anchors {
-            // Symmetric about the middle of the window, so the segment the
-            // anchors describe has the chart's own centre as its midpoint.
-            let offset = (anchor as f32 + 0.5) / anchors as f32 - 0.5;
-            let slot = ((first as f32 + visible as f32 * (0.5 + offset * DEMO_DRAFT_SPAN))
-                as usize)
-                .min(slots.saturating_sub(1));
-            let point = drawings::ChartPoint::at_time(
-                slot as f32 + 0.5,
-                centre + f64::from(offset) * band * DEMO_DRAFT_BAND_SPAN,
-                pane.slot_open_time(slot),
-            );
-            pane.drawings
-                .place_with(tool, &drawings::DrawingBand::Price, point, |tool| {
-                    drawings::NewDrawing {
-                        style: tool.default_style(),
-                        payload: tool.default_payload(),
-                    }
-                });
-        }
-        // The hand this run does not have. Parked on the line the anchors
-        // drew, which is where a real hand is sitting the instant a drag lets
-        // go — and where a channel used to be born with no width at all.
-        //
-        // `QUANTICK_DRAWING_CONSTRAIN=1` presses Shift for it. The levelled
-        // draft is a state of its own — a corridor held flat looks different
-        // from one following the pointer — and a modifier is the one input a
-        // capture run cannot supply any other way.
-        // With two anchors down the pointer belongs *on* the line they drew —
-        // the exact spot the reported defect lived at. With one, that spot is
-        // the anchor itself, so it steps aside far enough to leave a rubber
-        // band worth looking at.
-        let parked = if anchors >= 2 {
-            chart.center()
-        } else {
-            chart.center()
-                + egui::vec2(
-                    chart.width() * DEMO_DRAFT_POINTER_OFFSET.x,
-                    chart.height() * DEMO_DRAFT_POINTER_OFFSET.y,
-                )
-        };
-        pane.parked_hand = Some(pane::ParkedHand {
-            position: parked,
-            constrain: if constrain {
-                drawings::Constrain::Level
-            } else {
-                drawings::Constrain::Free
-            },
-        });
-    }
-
-    /// The `QUANTICK_VENUE_HISTORY_DEMO` hook: a venue candle prefix in front
-    /// of the bars cut from prints, delivered through the very path a feed's
-    /// reply takes, so what is photographed is the real seam and the real
-    /// loading state rather than a picture of them.
-    ///
-    /// `partial` stops one slice short: the prefix is installed and the run is
-    /// left open, which is the mid-load frame progressive delivery exists to
-    /// produce and the one no capture could otherwise catch.
-    fn apply_venue_history_demo(&mut self) {
-        /// Candles the venue-history scene installs: enough for the seam and
-        /// the divider to read, few enough to stay one screenful of context.
-        const DEMO_PREFIX_CANDLES: i64 = 90;
-        let Some(demo) = self.harness.venue_history_demo() else {
-            return;
-        };
-        let tab = self.active_tab_mut();
-        // Wait for bars to sit the prefix in front of: a seam needs both sides.
-        if tab.flow_pane.state.bars().len() < 12 {
-            return;
-        }
-        self.harness.venue_history_demo_staged();
-        let slice = match demo {
-            VenueHistoryDemo::Complete => quantick_feed::OhlcvSlice::Last { complete: true },
-            VenueHistoryDemo::Partial => quantick_feed::OhlcvSlice::More,
-        };
-        self.deliver_synthetic_prefix(DEMO_PREFIX_CANDLES, slice);
-    }
-
-    /// Deliver `candles` synthetic venue candles for the minutes immediately
-    /// before the first engine bar, so the prefix meets the tape without
-    /// overlapping it.
-    ///
-    /// The wiggle is a fixed function of the minute — the capture has to be
-    /// the same picture every run, or a visual diff means nothing. One
-    /// generator, shared by every hook that needs a prefix: a second one would
-    /// be a second history to keep honest.
-    /// Returns whether the candles were delivered: without a bar to sit them
-    /// in front of there is no seam to anchor on, and a caller that promised a
-    /// long history in its caption had better wait rather than photograph a
-    /// short one.
-    fn deliver_synthetic_prefix(&mut self, candles: i64, slice: quantick_feed::OhlcvSlice) -> bool {
-        let tab = self.active_tab_mut();
-        let Some(first) = tab.flow_pane.state.bars().first() else {
-            return false;
-        };
-        let (first_open, anchor) = (first.open_time, first.open);
-        let interval = quantick_feed::OHLCV_BASE_INTERVAL_MS;
-        let bars: Vec<quantick_engine::Bar> = (-candles..0)
-            .map(|minute| {
-                let open_time = first_open + minute * interval;
-                let drift = rust_decimal::Decimal::from(minute.rem_euclid(7) - 3);
-                let open = anchor + drift;
-                quantick_engine::Bar {
-                    open_time,
-                    close_time: open_time + interval - 1,
-                    open,
-                    high: open + rust_decimal::Decimal::from(2),
-                    low: open - rust_decimal::Decimal::from(2),
-                    close: open + rust_decimal::Decimal::from(minute.rem_euclid(3) - 1),
-                    buy_volume: rust_decimal::Decimal::from(2),
-                    sell_volume: rust_decimal::Decimal::from(3),
-                    trade_count: 7,
-                }
-            })
-            .collect();
-        tab.deliver_ohlcv_slice(interval, bars, slice);
-        true
-    }
-
     /// Play the alarm cues every tab's armed instances asked for this
     /// frame, and empty their queues.
     ///
@@ -7772,513 +6539,6 @@ impl QuantickApp {
         if self.apply_replay_action(ReplayAction::Control(ReplayControl::Restart)) {
             self.harness.replay_restart_taken();
         }
-    }
-
-    /// The `QUANTICK_STRATEGY_DEMO` hook: a named rectangle over the recent
-    /// tape with a force-bar instance armed on it (`1`), the arming dialog
-    /// open over it (`popup`), that dialog with the **alarm section
-    /// unfolded** (`alarm`), or an **alarm-only** instance wearing a
-    /// standing preview mark (`alarm-badge`). The rectangle spans the
-    /// visible middle of the chart so `QUANTICK_CONTEXT_MENU=chart`'s centre
-    /// click lands on it and opens the per-drawing menu. Consumed once the
-    /// chart has bars enough, like the drawings demo.
-    fn apply_strategy_demo(&mut self) {
-        let Some(mode) = self.harness.strategy_demo() else {
-            return;
-        };
-        /// Fewest bars before the demo stages: enough for the shipped
-        /// 20-body window to be warm and the rectangle to have tape to span.
-        const DEMO_STRATEGY_MIN_SLOTS: usize = 25;
-        /// How far back of the newest bar the rectangle's left edge sits.
-        const DEMO_STRATEGY_LOOKBACK_BARS: usize = 40;
-        /// How far past the newest bar its right edge reaches, keeping the
-        /// region alive into the future like a stretched hand-drawn one.
-        const DEMO_STRATEGY_AHEAD_BARS: f32 = 6.0;
-        /// Half-height of the region, as a fraction of the newest close.
-        const DEMO_STRATEGY_BAND_FRACTION: f64 = 0.03;
-        // The newest *closed* bar anchors the demo: `slots()` counts the
-        // forming partial too, whose `closed_bar` is `None` on almost every
-        // frame — bailing on it must keep the flag armed for the next
-        // frame, or the hook silently stages nothing.
-        let closed = self.active_tab_mut().flow_pane.closed_slots();
-        if closed < DEMO_STRATEGY_MIN_SLOTS {
-            return;
-        }
-        let Some(rectangle) = drawings::DRAWING_TOOLS
-            .into_iter()
-            .find(|tool| tool.id() == drawings::RECTANGLE_TOOL_ID)
-        else {
-            // No rectangle in the registry: staging can never succeed, so
-            // the flag is consumed rather than retried forever.
-            self.harness.strategy_demo_staged();
-            return;
-        };
-        let drawing_id = {
-            let pane = &mut self.active_tab_mut().flow_pane;
-            let newest = closed - 1;
-            let Some(close) = pane
-                .closed_bar(newest)
-                .and_then(|bar| rust_decimal::prelude::ToPrimitive::to_f64(&bar.close))
-            else {
-                return;
-            };
-            let start = newest.saturating_sub(DEMO_STRATEGY_LOOKBACK_BARS);
-            #[allow(clippy::cast_precision_loss)]
-            let anchors = [
-                drawings::ChartPoint::at_time(
-                    start as f32,
-                    close * (1.0 - DEMO_STRATEGY_BAND_FRACTION),
-                    pane.slot_open_time(start),
-                ),
-                // Past the newest bar no market time exists to name; the
-                // anchor carries none, like a hand-dropped one would.
-                drawings::ChartPoint::at_time(
-                    newest as f32 + DEMO_STRATEGY_AHEAD_BARS,
-                    close * (1.0 + DEMO_STRATEGY_BAND_FRACTION),
-                    None,
-                ),
-            ];
-            for point in anchors {
-                pane.drawings
-                    .place_with(rectangle, &drawings::DrawingBand::Price, point, |tool| {
-                        drawings::NewDrawing {
-                            style: drawings::DrawingStyle::default(),
-                            payload: tool.default_payload(),
-                        }
-                    });
-            }
-            let index = pane.drawings.items().len().saturating_sub(1);
-            pane.drawings.rename_at(index, "demo região");
-            pane.drawings.items()[index].id
-        };
-        // Staged: the rectangle exists, so the hook is consumed.
-        self.harness.strategy_demo_staged();
-        let mut form =
-            crate::strategy_presets::StoredPreset::starting_point(quantick_engine::Side::Buy);
-        // The alarm scenes tick the checkbox the trader would tick, and the
-        // share gate under it, so the section is unfolded with every control
-        // it owns on screen.
-        if matches!(
-            mode,
-            StrategyDemoMode::AlarmPopup
-                | StrategyDemoMode::AlarmSounds
-                | StrategyDemoMode::AlarmBadge
-        ) {
-            form.alarm = true;
-            form.alarm_when = "share".to_owned();
-            form.alarm_repeat = "cooldown".to_owned();
-            // A library clip with a cut, so the row that exists only for a
-            // clip is on screen too. The first standard clip, whatever it
-            // is called: a name here would go stale the day a clip is
-            // renamed, and the scene would silently photograph the
-            // system-sound caveat instead.
-            let clip = crate::audio::AlertSound::in_category(crate::audio::SoundCategory::Standard)
-                .next()
-                .expect("the shipped library has a standard clip");
-            form.alarm_sound = clip.token().to_owned();
-            form.alarm_play_secs = Some(crate::strategy_presets::DEFAULT_ALARM_PLAY_SECS);
-        }
-        match mode {
-            StrategyDemoMode::Armed => {
-                let _ = self.arm_strategy_instance(
-                    pane::PaneSide::Flow,
-                    drawing_id,
-                    &form,
-                    "demo BF".to_owned(),
-                );
-            }
-            StrategyDemoMode::AlarmBadge => {
-                form.alarm_only = true;
-                let _ = self.arm_strategy_instance(
-                    pane::PaneSide::Flow,
-                    drawing_id,
-                    &form,
-                    "demo alarm".to_owned(),
-                );
-                // Stand a provisional judgement on the badge. The mark is
-                // the surface under test, and the tape reaches it only when
-                // a force bar happens to be half-formed — so the scene
-                // stages the mark itself rather than waiting for a market
-                // that may not oblige before the shutter.
-                let pane = self.active_tab_mut().pane_mut(pane::PaneSide::Flow);
-                if let Some(instance) = pane.strategies.for_drawing_mut(drawing_id) {
-                    instance.mark = crate::strategy_anchors::AlarmMark::Preview;
-                }
-                // Placing a drawing selects it, and a selected drawing raises
-                // the context bar across its own top edge — which is where
-                // the badge this scene exists to photograph sits. Drop the
-                // selection so the badge is the thing on screen.
-                pane.drawings.select(None);
-            }
-            StrategyDemoMode::EndedBadge | StrategyDemoMode::PausedBadge => {
-                let _ = self.arm_strategy_instance(
-                    pane::PaneSide::Flow,
-                    drawing_id,
-                    &form,
-                    "demo BF".to_owned(),
-                );
-                let pane = self.active_tab_mut().pane_mut(pane::PaneSide::Flow);
-                if mode == StrategyDemoMode::EndedBadge {
-                    // End the span the way the trader does — by moving the
-                    // rectangle — rather than by stamping a state on. The
-                    // shared demo band reaches `DEMO_STRATEGY_AHEAD_BARS`
-                    // past the newest bar, which is why arming accepted it;
-                    // pulling both anchors behind the tape is the drag that
-                    // ends a region. Faking the state instead would
-                    // photograph the words over a band that can still fire,
-                    // and a reviewer would sign off on a scene the
-                    // application cannot produce.
-                    if let Some(index) = pane.drawings.index_of(drawing_id) {
-                        #[allow(clippy::cast_precision_loss)]
-                        let behind = closed.saturating_sub(2) as f32;
-                        for point in &mut pane.drawings.items_mut()[index].points {
-                            point.bar = point.bar.min(behind);
-                        }
-                    }
-                } else if let Some(index) = pane.drawings.index_of(drawing_id) {
-                    // A re-cut stranding an anchor is what sets this on a
-                    // real chart; nothing scripted can provoke one on cue.
-                    pane.drawings.items_mut()[index].off_series = true;
-                }
-                // As with the alarm badge: a selected drawing raises the
-                // context bar across the very edge the badge sits on.
-                pane.drawings.select(None);
-            }
-            StrategyDemoMode::Popup
-            | StrategyDemoMode::AlarmPopup
-            | StrategyDemoMode::AlarmSounds => {
-                if mode == StrategyDemoMode::AlarmSounds {
-                    self.surfaces.strategy_popup.stage_sound_picker();
-                }
-                let tab = self.active_tab().id;
-                self.surfaces
-                    .strategy_popup
-                    .open(tab, pane::PaneSide::Flow, drawing_id, form);
-            }
-        }
-        // This demo places its rectangle, which selects it, which closes a
-        // panel the launch may have asked for. Same door as the others.
-        self.carry_inspector_across_selection();
-    }
-
-    /// The `QUANTICK_FRVP_DEMO` hook: one fixed-range volume profile on the
-    /// flow pane. When the pane carries a venue history prefix the range
-    /// starts inside it, so the partial-coverage honesty label ("profile
-    /// from N of M bars") is on screen — the surface this hook exists to
-    /// photograph. Consumed once, like the drawings demo.
-    fn apply_frvp_demo(&mut self) {
-        /// One-minute venue candles the `=stress` scene installs behind the
-        /// tape — a time chart's worth of history, and far more than one fold
-        /// pass spends, so the range folds across frames rather than in the one
-        /// that placed it. The time pane folds them to whatever interval it is
-        /// showing, so the slot count is this number only at 1m — the scene is
-        /// about the fold surviving a long history, not about an exact count.
-        const FRVP_STRESS_CANDLES: i64 = 25_000;
-        // `=compare` places two adjacent profiles over the same stretch of
-        // map, one in each over-heatmap mode — the before/after of the
-        // silhouette decision in a single frame, which no toggle-and-wait
-        // pair of screenshots can prove as cleanly.
-        let Some(demo) = self.harness.frvp_demo() else {
-            return;
-        };
-        let FrvpDemo {
-            compare,
-            stress,
-            select,
-        } = demo;
-        // The venue's candles land on the **time** pane — that is the chart
-        // whose history runs to tens of thousands of bars, and the one a
-        // trader drops a session profile on. Every other scene stays on the
-        // flow pane, where the tape and the map are.
-        let side = if stress {
-            pane::PaneSide::Time(0)
-        } else {
-            pane::PaneSide::Flow
-        };
-        // The time pane is built the first time the split is shown, and
-        // `pane_mut` answers with the flow pane until it is. Waiting is the
-        // only honest option here: a stress scene photographed on the tape
-        // pane would be a picture of twenty bars captioned as twenty-five
-        // thousand.
-        if stress && self.active_tab_mut().time_panes.is_empty() {
-            return;
-        }
-        let slots = self.active_tab_mut().pane_mut(side).slots();
-        if slots < 12 {
-            return;
-        }
-        let Some(tool) = drawings::DRAWING_TOOLS
-            .into_iter()
-            .find(|tool| tool.id() == crate::frvp::TOOL_ID)
-        else {
-            self.harness.frvp_demo_placed();
-            return;
-        };
-        // `=stress` is the scene this object used to freeze the app on: a
-        // venue history longer than any single fold pass, with one profile
-        // over the whole of it. What it photographs is the *filling* state —
-        // a partial histogram and its `loading N of M bars` line — which is
-        // only a state at all because the fold is resumable. Pair it with
-        // QUANTICK_FRVP_FOLD_BUDGET=1 to hold that frame indefinitely.
-        //
-        // The delivery has to land before the range is placed: a scene that
-        // drew "the whole chart" over the handful of bars a pane starts with
-        // would be captioned as twenty-five thousand and be a picture of
-        // twenty. It stays armed and tries again next frame instead.
-        if stress
-            && !self.deliver_synthetic_prefix(
-                FRVP_STRESS_CANDLES,
-                quantick_feed::OhlcvSlice::Last { complete: true },
-            )
-        {
-            return;
-        }
-        self.harness.frvp_demo_placed();
-        // Re-read after the delivery: the prefix that just landed is part of
-        // the chart the range is about to span.
-        let slots = self.active_tab_mut().pane_mut(side).slots();
-        let prefix = self.active_tab_mut().pane_mut(side).history_prefix.len();
-        // The compare scene exists to photograph profiles *over the map*, and
-        // the map only covers bars closed after book capture began — the
-        // session's earliest bars never get cells. So it waits for enough
-        // freshly-built bars and anchors on those, where the coverage is.
-        if compare && slots.saturating_sub(prefix) < 60 {
-            self.harness.rearm_frvp_demo(demo);
-            return;
-        }
-        let pane = self.active_tab_mut().pane_mut(side);
-        // Straddle the seam when there is one; else the newest stretch.
-        let start = if prefix > 0 && prefix < slots {
-            prefix.saturating_sub(5)
-        } else {
-            slots.saturating_sub(30)
-        };
-        let end = (start + 29).min(slots - 1);
-        let close = pane
-            .closed_bar(end)
-            .and_then(|bar| rust_decimal::prelude::ToPrimitive::to_f64(&bar.close))
-            .unwrap_or(1.0);
-        let newest = slots - 1;
-        let ranges: &[(usize, usize, bool)] = if stress {
-            // The whole chart, oldest slot to newest: the range a trader
-            // drags when they want the session's own volume-by-price, and the
-            // one that used to take the window with it.
-            &[(0, newest, true)]
-        } else if compare {
-            // Two adjacent 25-bar ranges over the newest (map-covered) tape.
-            // Left object keeps the honest default; right one is forced to
-            // "always fill", the composed-into-the-map look under review.
-            &[
-                (newest.saturating_sub(49), newest.saturating_sub(25), true),
-                (newest.saturating_sub(24), newest, false),
-            ]
-        } else {
-            &[(start, end, true)]
-        };
-        // Selecting what the demo drew is what makes the *context bar* — the
-        // strip a trader edits a profile from — reachable without a click.
-        // Without it a validation run can photograph a profile but never the
-        // controls over it, which is exactly the surface whose placement this
-        // change is about (`ui-harness`: every surface owes a hook).
-        for &(from, to, outline) in ranges {
-            for slot in [from, to] {
-                pane.drawings.place_with(
-                    tool,
-                    &drawings::DrawingBand::Price,
-                    // On the slot's centre, where a trader's click lands —
-                    // anchoring on its trailing edge drew a box one candle
-                    // wider than the range it folded.
-                    drawings::ChartPoint::at_time(slot as f32, close, pane.slot_open_time(slot)),
-                    |tool| {
-                        let mut payload = tool.default_payload();
-                        if let Some(frvp) =
-                            payload.as_any_mut().downcast_mut::<drawings::FrvpPayload>()
-                        {
-                            frvp.outline_over_heatmap = outline;
-                        }
-                        drawings::NewDrawing {
-                            style: drawings::DrawingStyle::default(),
-                            payload,
-                        }
-                    },
-                );
-            }
-            if select {
-                let last = pane.drawings.items().len().saturating_sub(1);
-                pane.drawings.select(Some(last));
-            }
-        }
-        if select {
-            self.carry_inspector_across_selection();
-        }
-    }
-
-    /// The `QUANTICK_AVWAP_DEMO` hook: one anchored VWAP on the flow pane,
-    /// anchored a stretch back with the first two band pairs on — the band
-    /// stack, the fills and the anchor marker in a single deterministic
-    /// frame. Consumed once, like the drawings demo.
-    fn apply_avwap_demo(&mut self) {
-        /// Fewest bars the demo waits for before anchoring — enough for the
-        /// average and one band pair to visibly develop.
-        const DEMO_AVWAP_MIN_SLOTS: usize = 12;
-        /// How far back of the newest bar the demo drops its anchor.
-        const DEMO_AVWAP_LOOKBACK_BARS: usize = 40;
-        if !self.harness.avwap_demo() {
-            return;
-        }
-        let slots = self.active_tab_mut().flow_pane.slots();
-        if slots < DEMO_AVWAP_MIN_SLOTS {
-            return;
-        }
-        self.harness.avwap_demo_placed();
-        let Some(tool) = drawings::DRAWING_TOOLS
-            .into_iter()
-            .find(|tool| tool.id() == crate::avwap::TOOL_ID)
-        else {
-            return;
-        };
-        let pane = &mut self.active_tab_mut().flow_pane;
-        // Far enough back for the average to develop, on a bar that exists.
-        let anchor = slots
-            .saturating_sub(DEMO_AVWAP_LOOKBACK_BARS)
-            .min(slots - 1);
-        let close = pane
-            .closed_bar(anchor)
-            .and_then(|bar| rust_decimal::prelude::ToPrimitive::to_f64(&bar.close))
-            .unwrap_or(1.0);
-        pane.drawings.place_with(
-            tool,
-            &drawings::DrawingBand::Price,
-            drawings::ChartPoint::at_time(anchor as f32, close, pane.slot_open_time(anchor)),
-            |tool| {
-                let mut payload = tool.default_payload();
-                if let Some(avwap) = payload
-                    .as_any_mut()
-                    .downcast_mut::<drawings::AvwapPayload>()
-                {
-                    // 1σ ships on; the demo also opens 2σ so the stack and
-                    // its fill layering are on screen.
-                    avwap.bands[1].on = true;
-                }
-                drawings::NewDrawing {
-                    style: tool.default_style(),
-                    payload,
-                }
-            },
-        );
-        self.carry_inspector_across_selection();
-    }
-
-    /// Keep a `QUANTICK_DRAWING_INSPECTOR=1` request alive across a selection
-    /// the *app* made, rather than the trader.
-    ///
-    /// The context bar closes the panel on every selection change, which is
-    /// right for a trader clicking from one object to the next and wrong for a
-    /// demo hook that places objects and selects one on the frame it runs:
-    /// the pairing the harness table prescribes then photographs a chart with
-    /// no panel on it. Re-requested through `pending_open_settings`, the door
-    /// a tool that asks for its own settings on placement already uses, which
-    /// is applied *after* the clear rather than before it.
-    ///
-    /// One function rather than the line copied into each hook, because the
-    /// omission is silent: the demo that forgets it produces a screenshot that
-    /// looks merely uninteresting, and that is how three of these hooks came to
-    /// disagree about it.
-    fn carry_inspector_across_selection(&mut self) {
-        self.surfaces.drawing_chrome.carry_across_selection();
-    }
-
-    /// The `bands` half of the demo hook: on every indicator pane, a level on
-    /// the band's own value and a diagonal across it.
-    ///
-    /// Two objects, not seventeen: a pane is a fifth of the chart's height,
-    /// and a screenshot of every tool stacked in one would prove nothing
-    /// about the projection it exists to check. The level is placed *at a
-    /// value the series actually holds*, so a drawing that has drifted off
-    /// its curve is visible at a glance.
-    fn seed_band_demo(pane: &mut ChartPane, first: usize, visible: usize, slots: usize) {
-        let level_slot = (first + visible / 2).min(slots.saturating_sub(1));
-        let (left, right) = (
-            (first + visible / 8).min(slots.saturating_sub(1)),
-            (first + visible * 3 / 4).min(slots.saturating_sub(1)),
-        );
-        for (band, value) in pane.indicator_band_samples(level_slot) {
-            for tool in drawings::DRAWING_TOOLS {
-                let anchors: &[(usize, f64)] = match tool.id() {
-                    "horizontal-line" => &[(0, 0.0)],
-                    "trend-line" => &[(1, 0.0), (2, 0.0)],
-                    _ => continue,
-                };
-                for (which, _) in anchors {
-                    let (slot, value) = match which {
-                        // The level sits on the sampled value itself.
-                        0 => (level_slot, value),
-                        // The diagonal spans the window around it, so its
-                        // ends are inside the band without being on the curve.
-                        1 => (left, value * 0.5),
-                        _ => (right, value * 1.5),
-                    };
-                    let point = drawings::ChartPoint::at_time(
-                        slot as f32 + 0.5,
-                        value,
-                        pane.slot_open_time(slot),
-                    );
-                    pane.drawings
-                        .place_with(tool, &band, point, |tool| drawings::NewDrawing {
-                            style: drawings::DrawingStyle::default(),
-                            payload: tool.default_payload(),
-                        });
-                }
-            }
-        }
-    }
-
-    /// The `QUANTICK_DRAWINGS_DEMO_RECUT` hook: re-cut the bars under the demo
-    /// objects, so a screenshot shows what a timeframe switch does to them.
-    ///
-    /// It is the only way to reach the two surfaces this behaviour added
-    /// without a human touching the BARS selector: marks that survived a
-    /// re-cut and are still on their own instants, and a mark the new series
-    /// cannot reach, faded and labelled off-series. One extra object is placed
-    /// an hour before the first bar to produce the second.
-    fn apply_drawing_demo_recut(&mut self) {
-        if !self.harness.drawings_demo_recut() {
-            return;
-        }
-        let pane = &mut self.active_tab_mut().flow_pane;
-        // An anchor before anything the tab has loaded: honest input for the
-        // off-series path, not a flag set by hand.
-        if let Some(first) = pane.slot_open_time(0) {
-            let base = pane
-                .closed_bar(0)
-                .and_then(|bar| rust_decimal::prelude::ToPrimitive::to_f64(&bar.close))
-                .unwrap_or(1.0);
-            // A one-anchor tool by name, not `DRAWING_TOOLS[0]` — that is the
-            // trend line, which needs two, so a single `place_with` left a
-            // half-finished draft and no object at all. The whole point here
-            // is to produce one *completed* off-series mark.
-            let single_anchor = drawings::DRAWING_TOOLS
-                .into_iter()
-                .find(|tool| tool.id() == "horizontal-line");
-            if let Some(tool) = single_anchor {
-                let placed = pane.drawings.place_with(
-                    tool,
-                    &drawings::DrawingBand::Price,
-                    drawings::ChartPoint::at_time(0.5, base, Some(first - DEMO_OFF_SERIES_LEAD_MS)),
-                    |tool| drawings::NewDrawing {
-                        style: drawings::DrawingStyle::default(),
-                        payload: tool.default_payload(),
-                    },
-                );
-                debug_assert!(placed, "a horizontal line completes on one anchor");
-            }
-        }
-        // Half the bars, same trades — the plainest re-cut there is. Two
-        // settle frames because a spec change waits for the selector to hold
-        // still for one (`Tab::apply_spec_change`).
-        pane.tick_n = pane.tick_n.saturating_mul(2).max(2);
-        self.active_tab_mut().apply_spec_changes();
-        self.active_tab_mut().apply_spec_changes();
     }
 }
